webpki: rename client_verifier_builder -> client_verifier
This commit renames the `client_verifier_builder.rs` module to be named
`client_verifier.rs`. In a subsequent commit we'll move the verifier
impl that is built by the verifier builder into the same module. This
will make it easier to create a `server_verifier.rs` module for the
server certificate verifier equivalents, and keeps the builder and the
type it builds in the same module for clarity.
